events_config_consumer/pyproject.toml
[tool.poetry]
name = "event-logs-consumer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "event_config_consumer"}]

[tool.poetry.dependencies]
python = "^3.9"
aiokafka = "^0.8.0"
fastapi = "^0.105.0"
uvicorn = "^0.21.1"
gunicorn = "^20.1.0"
clickhouse-connect = "0.3.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
jsonpath-ng = "^1.6.1"
pandas = "^2.2.1"
pydantic-settings = "^2.2.1"
numpy = "^1.26.4"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.0"}



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
